Phase 1: extract core/render-config.js (marked + mermaid setup)
configureMarked / configureMermaid / getMermaidConfig + the FONT_FAMILY const
move out. Pure config (no callbacks into the app); imports marked/mermaid/hljs +
state. main.js ~1,688 -> ~1,639. Build + lint clean; 297 tests pass.
